refactor: modularize self-improvement engine
